Multiple bug fixes that previously went under the radar as well as changes to Teaching class
Teaching class no longer has the timeslot property as it can be derived
directly and easily from the location of the teaching in the schedule.
Implementing this change also fixed a bug in  random_walk. Instances of
the Teaching class can now also be initialised by another Teaching
object. This was done to allow for a neat way to fix the bugs in
best_teaching_swap (formerly just teaching_swap).